Refactor map UI and locations generally
