+++ packages/client/src/components/World/Scene.module.scss
