+++ packages/client/src/pages/map/MapInstanceChat.module.scss
@@ -1,0 +1,170 @@
+.iconCallChat{
+  display: flex;
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  margin: 5px 15px 20px 10px;
+  align-items: center;
+  color: #000;
+  z-index: 20;
+  &:hover {
+    cursor: pointer;
+  }
+  svg{
+    margin: 0 5px 0 0 ;
+  }
+  button {
+    height: 48px;
+  }
+}
+.instance-chat-container {
+  position: absolute;
+  bottom: 0;
+  display: flex;
+  flex-direction: row;
+  right: 0;
+  width: 40vw;
+  justify-content: flex-end;
+  z-index: 20;
+  margin: 5px 20px 20px 10px;
+  transition: all 225ms cubic-bezier(0, 0, 0.2, 1) 0ms;
+  // overflow: hidden;
+
+  &.messageContainerClosed{
+    width: 0;
+    overflow: hidden;
+  }
+
+  @media (max-width: 768px) {
+    margin: 0;
+    width: 100vw;
+  }
+
+  .chat-box {
+    display: flex;
+    flex-grow: 1;
+    align-items: stretch;
+    padding: 0px;
+    align-items: flex-end;
+    .chat-box:last-child {
+      padding-bottom: 0px;
+    }
+
+    .iconContainer{
+      display: flex;
+      height: inherit;
+      align-items: center;
+      padding: 0px 10px;
+      &:hover {
+        cursor: pointer;
+      }
+      svg{
+        margin: 10px 0;
+      }
+    }
+    .messageFieldContainer{
+      margin:0px;
+      .PrivateNotchedOutline-root-1{
+        border: 0px;
+      }
+
+      :global(.MuiInputLabel-formControl) {
+        transform: translate(0, 16px) scale(1);
+      }
+
+      :global(.MuiInput-root) {
+        margin-top: 8px;
+        margin-bottom: 7px;
+      }
+
+      :global(.MuiFormLabel-root.Mui-focused):not(.MuiFormLabel-filled),
+      :global(.MuiFormLabel-root.MuiFormLabel-filled):not(.Mui-focused){
+        display: none;
+      }
+
+      :global(.MuiInput-underline:before),
+      :global(.MuiInput-underline:after){
+        border-bottom: none
+      }
+
+    }
+
+    
+    :global(.MuiButton-root).iconContainerSend{
+      display: flex;
+      height: 37px;
+      align-items: center;
+      // padding: 5px 10px;
+      width: 37px;
+      margin: 5px;
+      min-width: 0px;
+      &:hover {
+        cursor: pointer;
+      }
+    }
+  }
+
+  .list-container {
+    display: flex;
+    flex-direction: column;
+    width: 50vw;
+
+    @media (max-width: 768px) {
+      width: 96vw;
+      margin: 20px;
+      height: 100%;
+    }
+
+
+  }
+
+  ul {
+    min-height: 0;
+  }
+  .message {
+    .userName{
+      font-weight: bold;
+    }
+    &.self {
+      & > div {
+        width: 100%;
+        display: flex;
+      }
+    }
+    &.other {
+      &:after {
+        content: '';
+        width: 40%;
+        display: block;
+      }
+
+      & > div {
+        width: 60%;
+        display: flex;
+      }
+    }
+  }
+}
+
+:global(.MuiPaper-root).message-wrapper{
+    display: flex;
+    flex-direction: column;
+    height: 30vh;   
+    background-color: transparent;
+    margin-left: -55px;
+    @media (max-width: 768px) {
+      margin-left: 0;
+    }
+    
+  .message-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    height: 100%;
+    background: transparent;
+    .first-message-placeholder {
+      font-size: 2em;
+      margin: auto;
+    }
+  }
+}
