+++ packages/client/src/pages/map/MapLayout.module.scss
@@ -1,0 +1,196 @@
+html {
+  overflow: hidden;
+  -webkit-user-select: none;
+  user-select: none;
+}
+* {
+  touch-action: none;
+}
+
+.container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.box {
+  display: flex;
+  width: 200px;
+  height: 150px;
+  margin: 10px;
+  border: solid black 2px;
+  text-align: center;
+
+  &:hover {
+    background-color: gray;
+    cursor: pointer;
+  }
+}
+
+.harmony-toggle {
+  position: absolute !important;
+  right: 0;
+  top: 6px !important;
+  display: flex;
+  align-items: center;
+  z-index: 20;
+  background-color: rgb(56, 56, 178);
+  height: 36px;
+  width: 36px;
+  color: white;
+  margin-right: 8px;
+  border-radius: 50%;
+
+  svg {
+    margin: 6px;
+  }
+
+  &:hover {
+    cursor: pointer;
+    background-color: lighten(rgb(56, 56, 178), 10%);
+  }
+}
+.fullScreen {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: white;
+  background: rgba(0, 0, 0, 0.2);
+  width: 50px;
+  height: 50px;
+  cursor: pointer;
+  position: fixed;
+  bottom: 20px;
+  left: 20px;
+  z-index: 20;
+  border-radius: 50%;
+  border: none;
+  padding: 0;
+  &::before {
+    content: ' ';
+    position: absolute;
+    width: 50px;
+    height: 50px;
+    display: block;
+    transition: all 0.15s cubic-bezier(0.18, 0.89, 0.32, 1.28);
+    background: #5f5ff1;
+    top: 0;
+    left: 0;
+    z-index: -1;
+    border-radius: 50%;
+    transform: scale(0);
+  }
+  &:hover {
+    &::before {
+      transform: scale(1);
+    }
+  }
+  @media (max-width: 450px) {
+    left: 10px;
+  }
+}
+
+.locationUserMenu {
+  display: inline-flex;
+  flex-direction: column;
+  margin: 20px 0;
+  padding: 0 20px;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 11;
+  max-height: calc(100vh - 120px);
+  overflow-y: auto;
+  min-width: 90px;
+  scrollbar-width: none; /* IE */
+  -ms-overflow-style: none; /* Firefox */
+  &::-webkit-scrollbar {
+    display: none; /* Safari and Chrome */
+  }
+
+  &::after {
+    content: ' ';
+    padding: 10px;
+    display: none;
+  }
+
+  @media (max-width: 768px) {
+    flex-direction: row;
+    top: 60px;
+    overflow-x: auto;
+    max-width: 100vw;
+    padding: 20px;
+    margin: 0;
+    &::after {
+      display: inline-block;
+    }
+  }
+  @media (max-width: 450px) {
+    padding-left: 0;
+    padding-right: 0;
+  }
+}
+
+.expandMenu {
+  width: 50px;
+  height: 50px;
+  border: none;
+  color: white;
+  background: rgba(0, 0, 0, 0.2);
+  padding: 0;
+  position: absolute;
+  top: 10px;
+  left: 30px;
+  border-radius: 50%;
+  z-index: 11;
+  svg {
+    width: 50%;
+    height: 50%;
+  }
+  &.expanded {
+    background: #5f5ff1;
+  }
+  &::before {
+    content: ' ';
+    position: absolute;
+    width: 50px;
+    height: 50px;
+    display: block;
+    transition: all 0.15s cubic-bezier(0.18, 0.89, 0.32, 1.28);
+    background: #5f5ff1;
+    top: 0;
+    left: 0;
+    z-index: -1;
+    border-radius: 50%;
+    transform: scale(0);
+  }
+  &:hover {
+    &::before {
+      transform: scale(1);
+    }
+  }
+  @media (min-width: 769px) {
+    display: none;
+  }
+  @media (max-width: 450px) {
+    width: 40px;
+    height: 40px;
+    left: 10px;
+    &::before {
+      width: 40px;
+      height: 40px;
+    }
+  }
+}
+
+.invite-toggle {
+  right: 0 !important;
+  top: 30px !important;
+  position: absolute !important;
+  z-index: 20;
+
+  button[aria-label='PersonAdd'] {
+    background: linear-gradient(30deg, #028000, #62ff00);
+    border: 2px solid white;
+  }
+}
